Closes #791, normalizes client buffer length type.
git-svn-id: http://developer.mangband.org/svn/mangband/trunk@994 802d84f3-b497-254e-bede-06c0a5a3456a

+++ src/client/netclient.c
@@ -520,7 +520,7 @@ int Net_input(void)
 
 int Flush_queue(void)
 {
-	short	len;
+	int	len;
 
 	if (!initialized) return 0;
